move function to more sensible file, sample 5k rows instead of 10k
